requirements.txt
requests
psutil
python-dotenv
APScheduler>=3.10
